Rename package name "perfectfitapp" to "sensorapi"
