+++ sensor_api/README.md
@@ -8,7 +8,7 @@ For developing it, then install the dependencies in `requirements-dev.txt`.
 # Run the server
 Example code:
 ```python
-from perfectfitapp.run import create_app
+from sensorapi.run import create_app
 host = '0.0.0.0'
 port = 8080
 app = create_app(host, port)
@@ -17,5 +17,5 @@ app.run(host=host, port=port, debug=True)
 
 # Test the server
 ```python
-pytest --cov perfectfitapp --cov-report term
+pytest --cov sensorapi --cov-report term
 ```
